Added images, Basic Blending capabilities + demonstration scene.
Blending demonstration scene demonstrates how draw ordering may result in effective transparency effects, it should be noted that this is not so simple for complex objects.

I have thus included a paper in the Resources folder which will be used in the future to extend the transparency capabilities of this engine.